Move location of PreparePageEventRequestOptions type
